+++ Cult of the Lamb/Rebirth/Helper.cs
@@ -1,11 +1,7 @@
 namespace Rebirth;
 
-[HarmonyPatch]
 public static class Helper
 {
-    
-    public static bool TooOld { get; set; }
-    
     public static bool IsOld(Follower follower)
     {
         if (Plugin.RebirthOldFollowers.Value)
@@ -17,7 +13,7 @@ public static class Helper
 
     public static bool DoHalfStats()
     {
-        return Random.Range(0f, 1f) <= 0.2f;
+        return Random.Range(0f, 1f) <= Plugin.XpPenaltyChance.Value / 100f;
     }
     
 }