+++ Cult of the Lamb/Rebirth/Patches.cs
@@ -10,19 +10,18 @@ public static class Patches
     {
         if (Victim.team == Health.Team.Team2)
         {
-            if (CustomItemManager.DropLoot(Plugin.RebirthItemInstance,0.05f))
+            if (Random.Range(0, 101) <= Plugin.EnemyDropRate.Value)
             {
                 Plugin.Log.LogInfo($"Got a Rebirth token from {__instance.name}!");
-                InventoryItem.Spawn(Plugin.RebirthItem, Random.Range(1, 3), __instance.transform.position);
+                InventoryItem.Spawn(Plugin.RebirthItem, Random.Range(Plugin.DropMinQuantity.Value, Plugin.DropMaxQuantity.Value + 1), __instance.transform.position);
             }
         }
-
-        if (Victim.name.ToLower(CultureInfo.InvariantCulture).Contains("breakable body pile"))
+        else if (Victim.name.ToLower(CultureInfo.InvariantCulture).Contains("breakable body pile"))
         {
-            if (CustomItemManager.DropLoot(Plugin.RebirthItemInstance,0.05f))
+            if (Random.Range(0, 101) <= Plugin.EnemyDropRate.Value)
             {
                 Plugin.Log.LogInfo($"Got a Rebirth token from {__instance.name}!");
-                InventoryItem.Spawn(Plugin.RebirthItem, Random.Range(1, 3), __instance.transform.position);
+                InventoryItem.Spawn(Plugin.RebirthItem, Random.Range(Plugin.DropMinQuantity.Value, Plugin.DropMaxQuantity.Value + 1), __instance.transform.position);
             }
         }
     }